Refactor camera movement into hook
