+++ src/components/bomb.tsx
@@ -1,30 +1,31 @@
-import { useGLTF } from "@react-three/drei";
-import { useFrame, type ThreeEvent } from "@react-three/fiber";
+import { useFrame, useThree } from "@react-three/fiber";
+import { useEffect, useRef, useState } from "react";
 import * as THREE from "three";
+import type { Indicator, Port } from "../generated/proto/bomb.pb";
+import { ModuleModuleType, type Module } from "../generated/proto/modules.pb";
+import { useModuleModel } from "../hooks/use-module-model";
 import { useGameStore } from "../hooks/use-game-store";
-import { getModuleRoot } from "../utils/node-finder";
+import BombProvider from "./bomb-provider";
 import {
   BigButtonModule,
   ClockModule,
+  KeypadModule,
   PasswordModule,
   SimonModule,
-  WiresModule,
-  KeypadModule,
   WhosOnFirstModule,
+  WiresModule,
 } from "./modules";
-import { ModuleModuleType, type Module } from "../generated/proto/modules.pb";
-import type { Indicator, Port } from "../generated/proto/bomb.pb";
-import BombProvider from "./bomb-provider";
 import MemoryModule from "./modules/memory-module";
 import MorseModule from "./modules/morse-module";
 import NeedyVentGasModule from "./modules/needy-vent-gas";
-import { useEffect, useRef, useState } from "react";
+import { positionToCoords } from "../utils/position-to-coords";
+import {
+  BOMB_HEIGHT,
+  CAMERA_DISTANCE_ZOOMED,
+  CAMERA_HEIGHT,
+} from "../utils/constants";
 
-const ZOOM_DISTANCE = 0.3;
-const PICKED_UP_HEIGHT = 0.2; // Height to lift the bomb when picked up
-const PICKUP_ANIMATION_DURATION = 500; // 0.5 seconds in milliseconds
-
-interface BombProps {
+interface Props {
   bombId?: string;
   modules?: { [key: string]: Module };
   startedAt?: number;
@@ -37,337 +38,292 @@ interface BombProps {
   indicators?: { [key: string]: Indicator };
 }
 
-export default function Bomb({ modules, startedAt, timerDuration }: BombProps) {
-  const { nodes, materials } = useGLTF("/bomb.glb") as any;
-  const { zoomToModule, selectedModuleId, reset } = useGameStore();
-  const [rotation, setRotation] = useState<[number, number, number]>([0, 0, 0]);
-  const [isDragging, setIsDragging] = useState(false);
-  const [isPickedUp, setIsPickedUp] = useState(false);
-  const [animatedHeight, setAnimatedHeight] = useState(0);
-  const [pickupStartTime, setPickupStartTime] = useState<number | null>(null);
-  const groupRef = useRef<THREE.Group>(null);
-  const previousMousePosition = useRef({ x: 0, y: 0 });
+function getClosestForwardRotationRadians(currentRotation: number): number {
+  const distTo0 = Math.abs(shortestAngularDelta(currentRotation, 0));
+  const distToPi = Math.abs(shortestAngularDelta(currentRotation, Math.PI));
 
-  // Animate the bomb pickup/putdown
-  useFrame(() => {
-    if (pickupStartTime !== null) {
-      const elapsed = Date.now() - pickupStartTime;
-      const progress = Math.min(elapsed / PICKUP_ANIMATION_DURATION, 1);
-      
-      // Use easing function for smoother animation
-      const easeOut = 1 - Math.pow(1 - progress, 3);
-      
-      // Determine start and end heights based on pickup state
-      const startHeight = isPickedUp ? 0 : PICKED_UP_HEIGHT;
-      const endHeight = isPickedUp ? PICKED_UP_HEIGHT : 0;
-      
-      setAnimatedHeight(startHeight + (endHeight - startHeight) * easeOut);
-      
-      if (progress >= 1) {
-        setPickupStartTime(null);
+  return distTo0 < distToPi ? 0 : Math.PI;
+}
+
+function shortestAngularDelta(from: number, to: number): number {
+  const TWO_PI = Math.PI * 2;
+  let delta = (to - from) % TWO_PI;
+  if (delta > Math.PI) delta -= TWO_PI;
+  if (delta < -Math.PI) delta += TWO_PI;
+  return delta;
+}
+
+function BombSimple({ modules, startedAt, timerDuration }: Props) {
+  const { camera, scene } = useThree();
+  const { nodes, materials } = useModuleModel("bomb");
+  const setZoomState = useGameStore((s) => s.setZoomState);
+  const {
+    zoomToModule,
+    selectedModuleId,
+    reset,
+    cameraLocked,
+    cameraTargetLookAt,
+    cameraTargetPosition,
+  } = useGameStore();
+
+  const [isPickedUp, setIsPickedUp] = useState(false);
+  const [isDragging, setIsDragging] = useState(false);
+
+  const [rotation, setRotation] = useState<[number, number, number]>([0, 0, 0]);
+  const [animatedHeight, setAnimatedHeight] = useState(0);
+
+  const lastPointerPosition = useRef<{ x: number; y: number } | null>(null);
+  const bombRef = useRef<THREE.Group>(null);
+  const raycaster = useRef(new THREE.Raycaster());
+  const pointer = useRef(new THREE.Vector2());
+
+  const defaultRotation: [number, number, number] = [0, 0, 0];
+
+  function isBombClicked(event: PointerEvent): boolean {
+    if (!bombRef.current) return false;
+
+    const rect = (event.target as Element).getBoundingClientRect();
+    const x = ((event.clientX - rect.left) / rect.width) * 2 - 1;
+    const y = -((event.clientY - rect.top) / rect.height) * 2 + 1;
+    pointer.current.set(x, y);
+
+    raycaster.current.setFromCamera(pointer.current, camera);
+    const intersects = raycaster.current.intersectObjects(scene.children, true);
+
+    const bombGroup = bombRef.current;
+    for (let i = 0; i < intersects.length; i++) {
+      const obj = intersects[i].object;
+      // If the object is the bomb or a child of the bomb group
+      if (obj === bombGroup || bombGroup.children.includes(obj)) {
+        return true;
       }
+    }
+    return false;
+  }
+
+  function isModuleClicked(event: PointerEvent): string | null {
+    if (!bombRef.current) return null;
+
+    const rect = (event.target as Element).getBoundingClientRect();
+    const x = ((event.clientX - rect.left) / rect.width) * 2 - 1;
+    const y = -((event.clientY - rect.top) / rect.height) * 2 + 1;
+    pointer.current.set(x, y);
+
+    raycaster.current.setFromCamera(pointer.current, camera);
+    const intersects = raycaster.current.intersectObjects(scene.children, true);
+
+    for (let i = 0; i < intersects.length; i++) {
+      const obj = intersects[i].object;
+      let parent = obj;
+      while (parent) {
+        if (parent.name === "BombCase") {
+          return null;
+        }
+
+        if (parent.userData && parent.userData.moduleId) {
+          return parent.userData.moduleId as string;
+        }
+        parent = parent.parent as THREE.Object3D;
+      }
+    }
+    return null;
+  }
+
+  useFrame(() => {
+    if (cameraLocked && cameraTargetPosition && cameraTargetLookAt) {
+      camera.position.lerp(cameraTargetPosition, 0.1);
+      camera.lookAt(cameraTargetLookAt);
+    }
+
+    const targetY = isPickedUp ? 1 : BOMB_HEIGHT;
+    const speed = 0.1;
+    setAnimatedHeight((prev) => prev + (targetY - prev) * speed);
+
+    if (!isPickedUp) {
+      setRotation((prev) => {
+        const targetY = getClosestForwardRotationRadians(prev[1]);
+
+        const rx = prev[0] + (defaultRotation[0] - prev[0]) * speed;
+        const ry = prev[1] + shortestAngularDelta(prev[1], targetY) * speed;
+        const rz = prev[2] + (defaultRotation[2] - prev[2]) * speed;
+
+        return [rx, ry, rz];
+      });
+    }
+
+    if (bombRef.current) {
+      bombRef.current.position.y = animatedHeight;
+      bombRef.current.rotation.x = rotation[0];
+      bombRef.current.rotation.y = rotation[1];
+      bombRef.current.rotation.z = rotation[2];
     }
   });
 
-  // Handle picking up the bomb
-  const handlePickUp = (e: ThreeEvent<MouseEvent>) => {
-    console.log("handlePickUp called");
-    if (!isPickedUp) {
-      console.log("Picking up bomb - starting animation");
-      e.stopPropagation();
-      setIsPickedUp(true);
-      setPickupStartTime(Date.now());
-      // Reset position to center of view when picked up
-      previousMousePosition.current.x = e.nativeEvent.clientX;
-      previousMousePosition.current.y = e.nativeEvent.clientY;
-    }
-  };
+  useEffect(() => {
+    function handlePointerDown(event: PointerEvent) {
+      // Bomb is already picked up
+      // 1) Check if user clicked on a module => rotate bomb forward & zoom in
+      if (isBombClicked(event)) {
+        if (isPickedUp) {
+          reset();
+        } else {
+          setIsPickedUp(true);
+        }
+      }
 
-  // Handle putting down the bomb
-  const handlePutDown = () => {
-    console.log("handlePutDown called");
-    if (isPickedUp) {
-      console.log("Putting down bomb - starting animation");
-      setIsPickedUp(false);
-      setPickupStartTime(Date.now());
-      // Reset rotation when putting down
-      setRotation([0, 0, 0]);
-      // Reset camera if in module view
-      if (selectedModuleId) {
-        console.log("Resetting camera from module view");
+      const modId = isModuleClicked(event);
+      if (modId) {
+        const module = modules?.[modId];
+        if (!isPickedUp) {
+          setIsPickedUp(true);
+          event.preventDefault();
+        }
+
+        const modulePosition = module?.position;
+        const { position: baseCoords } = positionToCoords(modulePosition!); // TODO: Handle undefined position / fix type
+        const closestRotationY = getClosestForwardRotationRadians(rotation[1]);
+        setRotation([defaultRotation[0], closestRotationY, defaultRotation[2]]);
+
+        const cameraTargetPosition = new THREE.Vector3(
+          baseCoords.x,
+          baseCoords.y + CAMERA_HEIGHT,
+          CAMERA_DISTANCE_ZOOMED,
+        );
+
+        const lookAt = new THREE.Vector3(
+          baseCoords.x,
+          cameraTargetPosition.y,
+          0,
+        );
+
+        zoomToModule(modId, cameraTargetPosition, lookAt);
+        event.preventDefault();
+        return;
+      }
+
+      // 2) If user clicked bomb => start dragging
+      if (isBombClicked(event)) {
+        setIsDragging(true);
+        lastPointerPosition.current = { x: event.clientX, y: event.clientY };
+        event.preventDefault();
+      }
+
+      // 3) Otherwise => clicked outside => put it down
+      else {
+        setIsPickedUp(false);
+        setIsDragging(false);
         reset();
       }
     }
-  };
 
-  // Handle global click to put down the bomb when clicking outside
-  useEffect(() => {
-    const handleGlobalClick = (e: MouseEvent) => {
-      console.log("Global click detected:", { 
-        isPickedUp, 
-        isDragging, 
-        selectedModuleId,
-        target: e.target
-      });
-      
-      if (isPickedUp && !isDragging) {
-        // Check if the click was on a canvas element
-        const canvasElement = document.querySelector("canvas");
-        if (!canvasElement) return;
+    function handlePointerMove(event: PointerEvent) {
+      if (isPickedUp && isDragging) {
+        if (lastPointerPosition.current) {
+          const dx = event.clientX - lastPointerPosition.current.x;
+          const dy = event.clientY - lastPointerPosition.current.y;
 
-        // If the click target is the canvas itself, handle accordingly
-        if (e.target === canvasElement) {
-          // If in module view, reset to picked up view
-          if (selectedModuleId) {
-            console.log("Resetting from module view to picked up view");
-            reset();
-          } else {
-            // Put the bomb down
-            console.log("Putting bomb down from global click");
-            handlePutDown();
-          }
+          const rotateSpeed = 0.01;
+          setRotation((prev) => {
+            let [rx, ry, rz] = prev;
+            rx += dy * rotateSpeed;
+            ry += dx * rotateSpeed;
+            return [rx, ry, rz];
+          });
         }
+
+        lastPointerPosition.current = { x: event.clientX, y: event.clientY };
       }
-    };
+    }
 
-    const handleGlobalPointerMove = (e: PointerEvent) => {
-      if (isDragging && isPickedUp) {
-        const deltaX = e.clientX - previousMousePosition.current.x;
-        const deltaY = e.clientY - previousMousePosition.current.y;
-
-        setRotation([
-          rotation[0] + deltaY * 0.01,
-          rotation[1] + deltaX * 0.01,
-          rotation[2],
-        ]);
-
-        previousMousePosition.current.x = e.clientX;
-        previousMousePosition.current.y = e.clientY;
-      }
-    };
-
-    const handleGlobalPointerUp = () => {
-      if (isDragging) {
+    function handlePointerUp() {
+      if (isPickedUp && isDragging) {
         setIsDragging(false);
       }
-    };
+    }
 
-    window.addEventListener("click", handleGlobalClick);
-    window.addEventListener("pointermove", handleGlobalPointerMove);
-    window.addEventListener("pointerup", handleGlobalPointerUp);
+    document.addEventListener("pointerdown", handlePointerDown);
+    document.addEventListener("pointermove", handlePointerMove);
+    document.addEventListener("pointerup", handlePointerUp);
 
     return () => {
-      window.removeEventListener("click", handleGlobalClick);
-      window.removeEventListener("pointermove", handleGlobalPointerMove);
-      window.removeEventListener("pointerup", handleGlobalPointerUp);
+      document.removeEventListener("pointerdown", handlePointerDown);
+      document.removeEventListener("pointermove", handlePointerMove);
+      document.removeEventListener("pointerup", handlePointerUp);
     };
-  }, [isPickedUp, isDragging, selectedModuleId, rotation, handlePutDown]);
-
-  // Handle mouse movement for rotation
-  const handlePointerDown = (e: ThreeEvent<PointerEvent>) => {
-    if (isPickedUp) {
-      e.stopPropagation();
-      setIsDragging(true);
-      previousMousePosition.current.x = e.nativeEvent.clientX;
-      previousMousePosition.current.y = e.nativeEvent.clientY;
-
-      // Capture the pointer to ensure we keep receiving events
-      // Using document.body for more reliable capture
-      document.body.setPointerCapture(e.nativeEvent.pointerId);
-    }
-  };
-
-  function onModuleClick(selected: THREE.Object3D[], event?: THREE.Event) {
-    console.log("onModuleClick called:", { 
-      selected: selected?.length, 
-      isPickedUp, 
-      selectedModuleId,
-      event: event?.type,
-      eventObject: event?.object
-    });
-    
-    if (!isPickedUp) {
-      console.log("Bomb not picked up, ignoring module click");
-      return;
-    }
-    
-    // Try to get the clicked object from the event first, then from selected array
-    let clickedObject: THREE.Object3D | null = null;
-    
-    if (event?.object) {
-      clickedObject = event.object;
-      console.log("Using event object:", clickedObject);
-    } else if (selected && selected[0]) {
-      clickedObject = selected[0];
-      console.log("Using selected object:", clickedObject);
-    }
-    
-    if (!clickedObject) {
-      console.log("No clicked object found");
-      return;
-    }
-
-    const module = getModuleRoot(clickedObject);
-    console.log("Module root found:", module);
-    console.log("Module userData:", module.userData);
-    
-    const moduleId = module.userData["moduleId"];
-    console.log("Module ID:", moduleId);
-    
-    if (!moduleId) {
-      console.log("No module ID found in userData");
-      return;
-    }
-    
-    // If already viewing this module, return to normal view
-    if (selectedModuleId === moduleId) {
-      console.log("Resetting from same module");
-      reset();
-      return;
-    }
-
-    const pos = module.getWorldPosition(new THREE.Vector3());
-    
-    // Calculate the world position of the bomb (parent group)
-    const bombCenter = new THREE.Vector3();
-    if (groupRef.current) {
-      groupRef.current.getWorldPosition(bombCenter);
-    }
-    
-    // Calculate direction from bomb center to module
-    const direction = pos.clone().sub(bombCenter).normalize();
-    
-    // Position camera slightly further back for better view
-    const cameraPosition = pos.clone().add(direction.multiplyScalar(ZOOM_DISTANCE));
-    
-    console.log("Zooming to module:", {
-      moduleId,
-      modulePos: pos,
-      bombCenter,
-      direction,
-      cameraPosition
-    });
-    
-    // Zoom to the module
-    zoomToModule(
-      moduleId,
-      cameraPosition,
-      pos // Look at the module position
-    );
-
-    // Stop event propagation to prevent bomb being put down
-    if (isDragging) {
-      setIsDragging(false);
-    }
-  }
+  }, [camera, scene, isPickedUp, isDragging, setZoomState, selectedModuleId]);
 
   return (
-    <>
-      <group
-        position={[0, 0.73 + animatedHeight, 0]}
-        rotation={isPickedUp ? rotation : [0, 0, 0]}
-        ref={groupRef}
+    <group ref={bombRef}>
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Case.geometry}
+        material={materials.Silver}
+        scale={[0.911, 1, 1]}
+        onPointerEnter={(e) => e.stopPropagation()}
+        name="BombCase"
       >
-        {/* The bomb shell with modules attached */}
-        <group
-          name="BombWithModules"
-          dispose={null}
-          onClick={isPickedUp ? undefined : handlePickUp}
-          onPointerDown={isPickedUp ? handlePointerDown : undefined}
-        >
-          {/* The bomb shell */}
+        <group position={[0, -BOMB_HEIGHT, 0]} scale={[1.098, 1, 1]}>
           <mesh
             castShadow
             receiveShadow
-            geometry={nodes.Case.geometry}
+            geometry={nodes.H.geometry}
             material={materials.Silver}
-            scale={[0.911, 1, 1]}
-          >
-            <group position={[0, -0.732, 0]} scale={[1.098, 1, 1]}>
-              <mesh
-                castShadow
-                receiveShadow
-                geometry={nodes.H.geometry}
-                material={materials.Silver}
-                position={[0, 0.73, 0]}
-              />
-              <mesh
-                castShadow
-                receiveShadow
-                geometry={nodes.H001.geometry}
-                material={materials.Silver}
-                position={[0, 0.73, 0]}
-              />
-              <mesh
-                castShadow
-                receiveShadow
-                geometry={nodes.H002.geometry}
-                material={materials.Silver}
-                position={[0.001, 0.73, 0]}
-              />
-              <mesh
-                castShadow
-                receiveShadow
-                geometry={nodes.V.geometry}
-                material={materials.Silver}
-                position={[0.005, 0.732, 0]}
-              />
-              <mesh
-                castShadow
-                receiveShadow
-                geometry={nodes.V003.geometry}
-                material={materials.Silver}
-                position={[-0.007, 0.732, 0]}
-              />
-            </group>
-            <mesh
-              castShadow
-              receiveShadow
-              geometry={nodes.Backplate.geometry}
-              material={materials.Silver}
-              scale={[1.098, 1, 1]}
-            />
-          </mesh>
-
-          {/* Modules */}
-          <BombProvider startedAt={startedAt} timerDuration={timerDuration}>
-            <Select
-              onChangePointerDown={(selected, event) => {
-                console.log("Select onChangePointerDown:", { 
-                  selected: selected.length, 
-                  event: event?.type 
-                });
-                if (selected.length > 0) {
-                  onModuleClick(selected, event);
-                }
-              }}
-              onChangePointerUp={(selected, event) => {
-                console.log("Select onChangePointerUp:", { 
-                  selected: selected.length, 
-                  event: event?.type 
-                });
-                if (selected.length > 0) {
-                  onModuleClick(selected, event);
-                }
-              }}
-              border={isPickedUp ? "1px solid #55aaff" : "none"}
-              backgroundColor={
-                isPickedUp ? "rgba(75, 160, 255, 0.1)" : "transparent"
-              }
-            >
-              {Object.entries(modules || {}).map(([moduleId, module]) => {
-                return renderModule(moduleId, module);
-              })}
-            </Select>
-          </BombProvider>
+            position={[0, BOMB_HEIGHT, 0]}
+          />
+          <mesh
+            castShadow
+            receiveShadow
+            geometry={nodes.H001.geometry}
+            material={materials.Silver}
+            position={[0, BOMB_HEIGHT, 0]}
+          />
+          <mesh
+            castShadow
+            receiveShadow
+            geometry={nodes.H002.geometry}
+            material={materials.Silver}
+            position={[0.001, BOMB_HEIGHT, 0]}
+          />
+          <mesh
+            castShadow
+            receiveShadow
+            geometry={nodes.V.geometry}
+            material={materials.Silver}
+            position={[0.005, BOMB_HEIGHT, 0]}
+          />
+          <mesh
+            castShadow
+            receiveShadow
+            geometry={nodes.V003.geometry}
+            material={materials.Silver}
+            position={[-0.007, BOMB_HEIGHT, 0]}
+          />
         </group>
-      </group>
-    </>
+        <mesh
+          castShadow
+          receiveShadow
+          geometry={nodes.Backplate.geometry}
+          material={materials.Silver}
+          scale={[1.098, 1, 1]}
+        />
+      </mesh>
+
+      <BombProvider startedAt={startedAt} timerDuration={timerDuration}>
+        {Object.entries(modules || {}).map(([moduleId, mod]) => (
+          <group key={moduleId} userData={{ moduleId }}>
+            {renderModule(moduleId, mod)}
+          </group>
+        ))}
+      </BombProvider>
+    </group>
   );
 }
 
 function renderModule(key?: string, module?: Module) {
-  switch (module?.type) {
+  if (!module) return null;
+
+  switch (module.type) {
     case ModuleModuleType.BIG_BUTTON:
       return (
         <BigButtonModule
@@ -461,3 +417,5 @@ function renderModule(key?: string, module?: Module) {
       return null;
   }
 }
+
+export default BombSimple;
