+++ src/components/scene.tsx
@@ -12,10 +12,11 @@ import * as THREE from "three";
 import type { Bomb as BombType } from "../generated/proto/bomb.pb";
 import { useGameStore } from "../hooks/use-game-store";
 import { GameService } from "../services/api";
-import Bomb from "./bomb-simple";
+import Bomb from "./bomb";
 import { useHighlight } from "./highlight-provider";
 import Table from "./table";
 import { CAMERA_DISTANCE, CAMERA_HEIGHT } from "../utils/constants";
+import { useCameraControl } from "../hooks/use-camera-controls";
 
 export default function Scene() {
   const { selected } = useHighlight();
@@ -24,6 +25,7 @@ export default function Scene() {
   const { cameraTargetPosition, cameraTargetLookAt } = useGameStore();
   const isSessionFetched = useRef<boolean>(false);
   const cameraRef = useRef<THREE.PerspectiveCamera>(null);
+  useCameraControl(cameraRef);
 
   // Default camera position and target
   const defaultCameraPosition = new THREE.Vector3(
